remove the contacts and rename to conversation
