+++ Sources/XMTPiOS/Dm.swift
@@ -69,14 +69,6 @@ public struct Dm: Identifiable, Equatable, Hashable {
 	}
 
 	public func updateConsentState(state: ConsentState) async throws {
-		if client.hasV2Client {
-			switch state {
-			case .allowed: try await client.contacts.allowGroups(groupIds: [id])
-			case .denied: try await client.contacts.denyGroups(groupIds: [id])
-			case .unknown: ()
-			}
-		}
-
 		try ffiConversation.updateConsentState(state: state.toFFI)
 	}
